Models

Description

This module defines the core data types and structures used throughout the application for representing chemical reactions, mechanisms, and interactants. It includes various identifiers, relationship types, and masks for data handling, as well as details on the process and health checks.

Exports

ReactionID, CatalystID, MoleculeID - Unique identifiers for reactions, catalysts, and molecules.

ACCELERATE, PRODUCT_FROM, REAGENT_IN - Relationship types that define connections between reactions and interactants.

Reaction, Catalyst, Molecule - Data types representing different chemical entities and their relationships.

Interactant, Explain - Types for handling interactants and mechanism explanations in the context of reactions.

NodeMask, PathMask, RawMechanismDetailsMask, RawReactionDetailsMask, RelMask - WEB masks to abstract away direct types from the Hasbolt library for better type safety and ease of use.

ReactionDetails, MechanismDetails, ProcessDetails - Data types that encapsulate detailed information about reactions and mechanisms.

ReactionNode, MechanismNode, RawReactionDetails, RawMechanismDetails - Nodes representing the raw data structures in the graph database.

MechanismID, StageID - Identifiers for mechanisms and stages within those mechanisms.

HealthCheck - Type used for performing health checks on the database connection.

Documentation

type ReactionID = Int

type CatalystID = Int

type MoleculeID = Int

data ACCELERATE

def - the default value for ACCELERATE corresponds to Standard Temperature and Pressure (STP): - temperature = 273.15 K (Kelvin) - pressure = 101.325 kPa (kilopascals)

Constructors

ACCELERATE

Fields

temperature :: [Float]

pressure :: [Float]

data Catalyst

Constructors

Catalyst

Fields

catalystId :: CatalystID

catalystSmiles :: String

catalystName :: Maybe String

data Molecule

Constructors

Molecule

Fields

moleculeId :: MoleculeID

moleculeSmiles :: String

moleculeIupacName :: String

newtype PRODUCT_FROM

Constructors

PRODUCT_FROM

Fields

productAmount :: Float

newtype REAGENT_IN

Constructors

REAGENT_IN

Fields

reagentAmount :: Float

data Reaction

Constructors

Reaction

Fields

reactionId :: ReactionID

reactionName :: String

data Interactant

Primary type for Reaction explanation

Constructors

IAccelerate ACCELERATE

ICatalyst Catalyst

IMolecule Molecule

IProductFrom PRODUCT_FROM

IReagentIn REAGENT_IN

IReaction Reaction

data Explain

Primary type for Mechanism explanation

Constructors

EMechanism Mechanism

EStage Stage

newtype NodeMask

Constructors

NodeMask

Fields

nodePropsMask :: Map Text Value

data PathMask

Constructors

PathMask

Fields

pathNodesMask :: [Interactant]

pathRelationshipsMask :: [Interactant]

pathSequenceMask :: [Int]

data RawMechanismDetailsMask

Constructors

RawMechanismDetailsMask

Fields

rawMechanismMask :: NodeMask

rawContextMask :: RelMask

rawStagesMask :: [NodeMask]

rawIncludeMask :: [RelMask]

rawParticipantsMask :: [NodeMask]

data RawReactionDetailsMask

Constructors

RawReactionDetailsMask

Fields

rawReactionMask :: NodeMask

rawReagentsMask :: [NodeMask]

rawProductsMask :: [NodeMask]

rawInboundMask :: [RelMask]

rawOutboundMask :: [RelMask]

rawAccelerateMask :: [RelMask]

rawCatalystsMask :: [NodeMask]

newtype RelMask

Constructors

RelMask

Fields

relPropsMask :: Map Text Value

data ReactionDetails

Constructors

ReactionDetails

Fields

reaction :: Reaction

inboundReagents :: [(REAGENT_IN, Molecule)]

outboundProducts :: [(PRODUCT_FROM, Molecule)]

conditions :: [(ACCELERATE, Catalyst)]

data MechanismDetails

Constructors

MechanismDetails

Fields

mechanismContext :: (Mechanism, FOLLOW)

stageInteractants :: [(Stage, [Interactant])]

data ProcessDetails

Constructors

ProcessDetails

Fields

reactionDetails :: ReactionDetails

mechanismDetails :: MechanismDetails

type ReactionNode = Node

Represents a node in a reaction graph

type MechanismNode = Node

Represents a node in a mechanism graph

data RawReactionDetails

Constructors

RawReactionDetails

Fields

rawReaction :: Node

rawReagents :: [Node]

rawProducts :: [Node]

rawInbound :: [Relationship]

rawOutbound :: [Relationship]

rawAccelerate :: [Relationship]

rawCatalysts :: [Node]

data RawMechanismDetails

Constructors

RawMechanismDetails

Fields

rawMechanism :: Node

rawInteractants :: [Node]

rawInclude :: [Relationship]

rawStages :: [Node]

rawFollow :: Relationship

type MechanismID = Int

type StageID = Int

data Mechanism

Constructors

Mechanism

Fields

mechanismId :: MechanismID

mechanismName :: String

mechanismType :: String

mechanismActivationEnergy :: Float

newtype FOLLOW

Constructors

FOLLOW

Fields

description :: String

data Stage

Constructors

Stage

Fields

stageOrder :: StageID

stageName :: String

stageDescription :: String

stageProducts :: [String]

data INCLUDE

Constructors

INCLUDE

data HealthCheck

Constructors

HealthCheck

Fields

status :: String

message :: String